+++ ER4Manager/ER4Manager_1/Contacts.aspx.cs
@@ -77,7 +77,7 @@ public partial class Contacts : System.Web.UI.Page
         DataTable dt = new DataTable();
         System.Data.DataRow newrow;
 
-        dt.Columns.Add(new DataColumn("CONTACT_ID", typeof(Int16)));
+        dt.Columns.Add(new DataColumn("CONTACT_ID", typeof(int)));
         dt.Columns.Add(new DataColumn("OLD_CONTACT_ID", typeof(string)));
         dt.Columns.Add(new DataColumn("EMAIL", typeof(string)));
         dt.Columns.Add(new DataColumn("CONTACT_NAME", typeof(string)));
@@ -131,7 +131,7 @@ public partial class Contacts : System.Web.UI.Page
         DataTable dt = new DataTable();
         System.Data.DataRow newrow;
 
-        dt.Columns.Add(new DataColumn("CONTACT_ID", typeof(Int16)));
+        dt.Columns.Add(new DataColumn("CONTACT_ID", typeof(int)));
         dt.Columns.Add(new DataColumn("OLD_CONTACT_ID", typeof(string)));
         dt.Columns.Add(new DataColumn("EMAIL", typeof(string)));
         dt.Columns.Add(new DataColumn("CONTACT_NAME", typeof(string)));
